splitting up envelopes
