+++ src/components/envelopes/EnvelopeOne.js
@@ -1,0 +1,49 @@
+import EnvelopeBasic from "./envelope-images/Envelope.png";
+import EnvelopeGlow from "./envelope-images/EnvelopeGlow.png";
+import React from "react";
+import { useState } from 'react';
+import SpeechBubbleReuse from "../speech-bubble/SpeechBubbleReuse";
+import Letter from "../letter/Letter";
+import "./envelope.css";
+
+const EnvelopeOne = ({letterTwo}) => {
+  const [displayingLetter, showLetter] = useState(false);
+  const [speechBubble, setSpeechbubble] = useState(false);
+
+  const bubbleText = () => {
+    return(
+      "We have two main suspects, Andy & Tiffany... There is another letter there, read it."
+    )
+  }
+
+  const letterContent = () => {
+    return(
+      <div>
+        <p>Dear Andy,</p>
+        <p>This is letter one</p>
+        <p>Wishing you the best,</p>
+        <p>Tiffany</p>
+      </div>
+    );
+  }
+
+  const clickHandler = () => {
+    showLetter(true);
+    setSpeechbubble(true);
+    letterTwo(true);
+
+  }
+
+  return(
+    <div>
+      <img className="envelopeOutline" onClick={() => clickHandler()} src={EnvelopeBasic} alt="envelope" onMouseOver={e => e.currentTarget.src = EnvelopeGlow } onMouseOut={e => e.currentTarget.src = EnvelopeBasic }/>
+      <Letter displayingLetter={displayingLetter} showLetter={showLetter}>
+        {letterContent()}
+      </Letter>
+      <SpeechBubbleReuse display={speechBubble} showBubble={setSpeechbubble} words={bubbleText}>
+      </SpeechBubbleReuse>
+    </div>
+  );
+};
+
+export default EnvelopeOne;
