+++ src/components/envelopes/LetterFourContent.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+const LetterFourContent = () => { 
+
+  return(
+    <div>
+      <p>Yo Priya,</p>
+      <p>You owe me 69 dino bucks</p>
+      <p>give me now or else</p>
+      <p>Tiffany x</p>
+    </div>
+  );
+};
+
+export default LetterFourContent;
