+++ src/components/envelopes/LetterSixContent.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+const LetterSixContent = () => { 
+
+  return(
+    <div>
+      <p>Hello sexy dinoboy,</p>
+      <p>I love your hat</p>
+      <p>pls cum and put me in cuffs ;) </p>
+      <p>Priya xoxox</p>
+    </div>
+  );
+};
+
+export default LetterSixContent;
