+++ src/components/envelopes/LetterTwoContent.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+const LetterTwoContent = () => { 
+
+  return(
+    <div>
+      <p>Dear Diary,</p>
+      <p>Tiffany smells</p>
+      <p>Want divorse</p>
+      <p>Andy</p>
+    </div>
+  );
+};
+
+export default LetterTwoContent;
