+++ src/components/envelopes/envelope.css
@@ -1,0 +1,10 @@
+.envelopeOutline{
+  height: 3%;
+  width: 4.5%;
+  position: absolute;
+  top: 70%;
+  right: 65%;
+  transform: rotate(0deg) skewX(-50deg) skewY(0deg);
+  border: none;
+  outline: transparent;
+}
